+++ event-buffer/event-buffer-core/src/test/java/uk/gov/justice/services/event/buffer/core/service/ConsecutiveEventBufferServiceTest.java
@@ -30,7 +30,6 @@ import java.time.ZonedDateTime;
 import java.util.UUID;
 import java.util.stream.Stream;
 
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InOrder;