+++ event-buffer/event-buffer-liquibase/liquibase.sh
@@ -1,15 +1,18 @@
 #!/usr/bin/env bash
 
 CONTEXT_NAME=framework
-EVENT_STORE_VERSION=11.0.0-M15-SNAPSHOT
+EVENT_STORE_VERSION=17.102.0-SNAPSHOT
 
 #fail script on error
 set -e
 
+LIQUIBASE_COMMAND=update
+#LIQUIBASE_COMMAND=dropAll
+
 
 function runEventBufferLiquibase() {
     echo "running EventBufferLiquibase"
-    java -jar target/event-buffer-liquibase-${EVENT_STORE_VERSION}.jar --url=jdbc:postgresql://localhost:5432/${CONTEXT_NAME}viewstore --username=${CONTEXT_NAME} --password=${CONTEXT_NAME} --logLevel=info update
+    java -jar target/event-buffer-liquibase-${EVENT_STORE_VERSION}.jar --url=jdbc:postgresql://localhost:5432/${CONTEXT_NAME}viewstore --username=${CONTEXT_NAME} --password=${CONTEXT_NAME} --logLevel=info ${LIQUIBASE_COMMAND}
     echo "Finished EventBufferLiquibase"
 }