+++ event-sourcing/subscription-manager/src/main/java/uk/gov/justice/services/event/sourcing/subscription/error/StreamProcessingException.java
@@ -1,0 +1,8 @@
+package uk.gov.justice.services.event.sourcing.subscription.error;
+
+public class StreamProcessingException extends RuntimeException{
+
+    public StreamProcessingException(final String message, final Throwable cause) {
+        super(message, cause);
+    }
+}