+++ event-store-management/event-store-management-command-handler-extension/src/test/java/uk/gov/justice/services/eventstore/management/publishing/EnablePublishingCommandHandlerTest.java
@@ -6,7 +6,6 @@ import static uk.gov.justice.services.jmx.api.parameters.JmxCommandRuntimeParame
 
 import uk.gov.justice.services.eventstore.management.commands.DisablePublishingCommand;
 import uk.gov.justice.services.eventstore.management.commands.EnablePublishingCommand;
-import uk.gov.justice.services.jmx.api.parameters.JmxCommandRuntimeParameters;
 
 import java.util.UUID;