+++ event-store-management/event-store-management-core/src/test/java/uk/gov/justice/services/eventstore/management/aggregate/snapshot/regeneration/commands/RebuildSnapshotCommandTest.java
@@ -3,7 +3,6 @@ package uk.gov.justice.services.eventstore.management.aggregate.snapshot.regener
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 
-import org.hamcrest.CoreMatchers;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;